+++ src/main/java/org/joelson/cts/calculator/model/builder/GardenBuilder.java
@@ -47,7 +47,7 @@ public class GardenBuilder {
     public GeneratorBuilder createGenerator(
             String name, Amount baseCost, float compoundingCost, Amount baseProduction, int baseChargeTime) {
         Generator generator = new Generator(name, baseCost.multiplyBy(costMultiplier), compoundingCost, baseProduction,
-                baseChargeTime / productionMultiplier);
+                baseChargeTime / (float) productionMultiplier);
         garden.addGenerator(generator);
         return new GeneratorBuilder(this, generator);
     }
